PTR-01-04: Added Team based tags, and other color customization
You can now add color to the name and message of the chat. Also has team
based colors too, so if you want to have a different colored name and
still want to have the ability to have them team colored you can now!
Either use it to confuse players of what team they are on, etc.